transaction_monitoring/requirements/base.txt
# Base requirements for Transaction Monitoring and Fraud Detection System

# Django and extensions
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
django-filter>=23.1,<24.0
django-crispy-forms>=2.0,<3.0
django-bootstrap5>=23.1,<24.0
django-cors-headers>=4.0.0,<5.0.0
django-redis>=5.2.0,<6.0.0

# Database
psycopg2-binary>=2.9.5,<3.0.0

# Celery
celery>=5.2.7,<6.0.0
redis>=4.5.1,<5.0.0

# Authentication
djangorestframework-simplejwt>=5.2.2,<6.0.0

# Utilities
python-dateutil>=2.8.2,<3.0.0
pytz>=2022.7,<2023.0
requests>=2.28.2,<3.0.0
pyyaml>=6.0,<7.0
networkx>=3.0,<4.0

# Machine Learning
scikit-learn>=1.2.2,<2.0.0
pandas>=1.5.3,<2.0.0
numpy>=1.24.2,<2.0.0

# Visualization
matplotlib>=3.7.1,<4.0.0
seaborn>=0.12.2,<0.13.0